force both to float so that there is no shadowing

+++ python/test_convolve.py
@@ -4,8 +4,8 @@
 import scipy
 import loudia
 
-a = scipy.array([1, 2, 3, 4, 5])
-b = scipy.array([2, 4, 5])
+a = scipy.array([1, 2, 3, 4, 5], dtype = 'f4')
+b = scipy.array([2, 4, 5], dtype = 'f4')
 
 rr = loudia.convolve( a, b )
 rs = scipy.convolve( a, b )
